+++ api/src/main/java/net/jqwik/api/arbitraries/LongArbitrary.java
@@ -10,7 +10,7 @@ import static org.apiguardian.api.API.Status.*;
  * Fluent interface to configure the generation of Long and long values.
  */
 @API(status = MAINTAINED, since = "1.0")
-public interface LongArbitrary extends Arbitrary<Long> {
+public interface LongArbitrary extends NumericalArbitrary<Long, LongArbitrary> {
 
 	/**
 	 * Set the allowed lower {@code min} (included) and upper {@code max} (included) bounds of generated numbers.