+++ engine/src/main/java/net/jqwik/engine/properties/arbitraries/DefaultBigDecimalArbitrary.java
@@ -35,6 +35,13 @@ public class DefaultBigDecimalArbitrary extends AbstractArbitraryBase implements
 	}
 
 	@Override
+	public BigDecimalArbitrary withDistribution(final RandomDistribution distribution) {
+		DefaultBigDecimalArbitrary clone = typedClone();
+		clone.generatingArbitrary.distribution = distribution;
+		return clone;
+	}
+
+	@Override
 	public BigDecimalArbitrary between(BigDecimal min, boolean minIncluded, BigDecimal max, boolean maxIncluded) {
 		min = (min == null) ? DEFAULT_MIN : min;
 		max = (max == null) ? DEFAULT_MAX : max;